Path to example breath test data file

Path to example breath test data file

btcore_file(filename = NULL, full.names = FALSE)

Arguments

filename
example file in extdata directory without path. Case sensitive on Unix. When filename is missing, returns the names of the available sample files.

full.names
When filename is NULL, return full path names

Value

full filename to example file to use in read_xxx

Examples

btcore_file,Path to example breath test data file,Path to example breath test data file — btcore_file,Source: R/btcore_file.R,btcore_file.Rd,character(0),btcore_file,character(0),list(title = "Description", contents = "
Path to example breath test data file
"),list(description = "Path to example breath test data file"),btcore_file(filename = NULL, full.names = FALSE),list(list(name = "filename", description = "
example file in extdata directory without path. \nCase sensitive on Unix. When filename is missing, returns the names of the \navailable sample files.
"), list(name = "full.names", description = "
When filename is NULL, return full path names
")),TRUE,
  head(btcore_file())

#> [1] "350_20023_0_GERWithNan.txt"       "350_20043_0_GER.txt"             
#> [3] "350_20043_0_GERBadHeader.txt"     "350_20043_0_GERDuplicateTime.txt"
#> [5] "350_20043_0_GERNoData.txt"        "350_20043_0_GERNoT50.txt"        

  filename = btcore_file("IrisMulti.TXT")
  data = read_iris(filename)

,list(list(title = "Value", contents = "
full filename to example file to use in read_xxx
", slug = "value")),2021,list(name = "breathtestcore", version = "0.8.0.9001", authors = "Dieter Menne, Benjamin Misselwitz, Mark Fox, University Hospital of Zurich, Dep. Gastroenterology"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "breathtestcore"),list(css = "C:/Users/Dieter/Documents/RPackages/breathtestcore/pkgdown/extra.css"),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Reference\n
\n
\n  Articles\n
", right = "
\n  \n    \n     \n  \n
"),list(bootswatch = "cosmo", .present = TRUE),list(version = "1.6.1"),FALSE