Broom-style tidying methods for breathtestfit

Broom-method tidy to streamline the results of class breathttestfit as generated by nls_fit or nlme_fit. Returns the fit coefficients and half-emptying time t50 with the Maes/Ghoos method; additional parameters should be extracted with coef.

# S3 method for breathtestfit
tidy(x, ...)

Arguments

x
Object of class breathttestfit

...
other parameters passed to methods

Value

A tibble/data frame with columns

patient_id

Patient Id (character)

group

Treatment or patient group (character)

m

Fraction metabolized

k

Time constant (1/minutes)

beta

The so-called lag parameters, no dimension

t50

Emptying half time in minutes as calculated following Maes/Ghoos

See also

Examples

library(broom)
# Generate simulated data
data = cleanup_data(simulate_breathtest_data()$data)
# Fit with the population method
fit = nlme_fit(data)
# Output coefficients
tidy(fit)

#> # A tibble: 10 x 6
#>    patient_id group     m       k  beta   t50
#>    <chr>      <chr> <dbl>   <dbl> <dbl> <dbl>
#>  1 rec_01     A      29.1 0.0138   1.60  76.0
#>  2 rec_02     A      23.5 0.00982  2.20 133. 
#>  3 rec_03     A      36.7 0.00732  1.47 134. 
#>  4 rec_04     A      38.3 0.00728  1.51 137. 
#>  5 rec_05     A      29.2 0.0125   2.63 117. 
#>  6 rec_06     A      65.1 0.00570  1.98 214. 
#>  7 rec_07     A      41.5 0.00822  2.29 163. 
#>  8 rec_08     A      40.2 0.0110   2.10 115. 
#>  9 rec_09     A      42.8 0.0120   2.31 113. 
#> 10 rec_10     A      75.9 0.00648  1.54 156. 

# All coefficients in the long form
coef(fit)

#> # A tibble: 80 x 5
#>    patient_id group parameter method                      value
#>    <chr>      <chr> <chr>     <chr>                       <dbl>
#>  1 rec_01     A     m         exp_beta                 29.1    
#>  2 rec_01     A     k         exp_beta                  0.0138 
#>  3 rec_01     A     beta      exp_beta                  1.60   
#>  4 rec_01     A     t50       bluck_coward             12.8    
#>  5 rec_01     A     t50       maes_ghoos               76.0    
#>  6 rec_01     A     t50       maes_ghoos_scintigraphy   8.89   
#>  7 rec_01     A     tlag      bluck_coward            -16.0    
#>  8 rec_01     A     tlag      maes_ghoos               34.3    
#>  9 rec_02     A     m         exp_beta                 23.5    
#> 10 rec_02     A     k         exp_beta                  0.00982
#> # ... with 70 more rows