Augmented prediction for breathtest fit

Broom-methods to compute predicted values from the results of class breathttestfit as generated by breathtestcore::nls_fit or breathtestcore::nlme_fit.

# S3 method for breathtestfit
augment(x, by = NULL, minute = NULL, dose = 100,
  ...)

Arguments

x
Object of class breathttestfit

by
When by is NULL, predictions for the original data values are returned. When by is a positive number, it is used as a step size for a sequence of minutes from 0 to the maximum value of minute in data set.

minute
When a vector is passed, this overrides settings in by, and predictions are calculated at the requested minute values.

dose
13C acetate or octanoate dose

...
other parameters passed to methods

Value

When by is NULL, returns one row for each original observation pdr, and column fitted. If new data are given, i.e. when one of parameter by or minute is not null, only column fitted is added.

Examples

library(broom)
data = cleanup_data(simulate_breathtest_data()$data)
fit = nls_fit(data)
augment(fit, minute = c(60, 120))

#> Source: local data frame [20 x 4]
#> Groups: <by row>
#> 
#> # A tibble: 20 × 4
#>    patient_id group minute    fitted
#> *       <chr> <chr>  <dbl>     <dbl>
#> 1      rec_01     A     60 26.120375
#> 2      rec_01     A    120 16.130834
#> 3      rec_02     A     60 15.748625
#> 4      rec_02     A    120 17.048118
#> 5      rec_03     A     60  7.023930
#> 6      rec_03     A    120  5.551014
#> 7      rec_04     A     60 28.951545
#> 8      rec_04     A    120 21.769491
#> 9      rec_05     A     60 19.920000
#> 10     rec_05     A    120 22.853289
#> 11     rec_06     A     60 18.788741
#> 12     rec_06     A    120 18.216992
#> 13     rec_07     A     60 19.618450
#> 14     rec_07     A    120 10.813736
#> 15     rec_08     A     60 19.079337
#> 16     rec_08     A    120 17.137230
#> 17     rec_09     A     60 14.429243
#> 18     rec_09     A    120 13.456638
#> 19     rec_10     A     60 16.951130
#> 20     rec_10     A    120 13.580971