Augmented prediction for breathtest fit

Broom method augment to compute predicted values from the results of class breathttestfit as generated by nls_fit or nlme_fit.

# S3 method for breathtestfit
augment(x, by = NULL, minute = NULL, dose = 100, ...)

Arguments

x
Object of class breathttestfit

by
When by is NULL, predictions for the original data values are returned. When by is a positive number, it is used as a step size for a sequence of minutes from 0 to the maximum value of minute in data set.

minute
When a vector is passed, this overrides settings in by, and predictions are calculated at the requested minute values.

dose
13C acetate or octanoate dose

...
other parameters passed to methods

Value

When by is NULL, returns one row for each original observation pdr, and column fitted. If new data are given, i.e. when one of parameter by or minute is not null, only column fitted is added.

See also

Examples

augment.breathtestfit,Augmented prediction for breathtest fit,Augmented prediction for breathtest fit — augment.breathtestfit,Source: R/breathtestfit_broom.R,augment.breathtestfit.Rd,character(0),augment.breathtestfit,character(0),list(title = "Description", contents = "
Broom method augment to compute predicted values \nfrom  the results of class breathttestfit  as generated by  \nnls_fit or nlme_fit.
"),list(description = "Broom method augment to compute predicted values \nfrom  the results of class breathttestfit  as generated by  \nnls_fit or nlme_fit."),# S3 method for breathtestfit
augment(x, by = NULL, minute = NULL, dose = 100, ...),list(list(name = "x", description = "
Object of class breathttestfit
"), list(name = "by", description = "
When by is NULL, predictions for the original data values\nare returned.  When by is a positive number, it is used as a step \nsize for a sequence of minutes from 0 to the maximum value of minute in data set.
"), list(name = "minute", description = "
When a vector is passed, this overrides settings in by,\nand predictions are calculated at the requested minute values.
"), 
    list(name = "dose", description = "
13C acetate or octanoate dose
"), list(name = "...", description = "
other parameters passed to methods
")),TRUE,
library(broom)
# Generate simulated data
data = cleanup_data(simulate_breathtest_data(n_records = 3)$data)
# Fit using the curves individually
fit = nls_fit(data)
# Predict values at t=60 and t=120
augment(fit, minute = c(60, 120))

#> # A tibble: 6 x 4
#> # Rowwise: 
#>   patient_id group minute fitted
#>   <chr>      <chr>  <dbl>  <dbl>
#> 1 rec_01     A         60  14.4 
#> 2 rec_01     A        120   8.61
#> 3 rec_02     A         60  25.6 
#> 4 rec_02     A        120  18.1 
#> 5 rec_03     A         60  10.4 
#> 6 rec_03     A        120   6.33

,list(list(title = "Value", contents = "
When by is NULL,  returns one row for each \noriginal observation pdr, and column fitted. If new data are given, \ni.e. when one of parameter by or minute is not null, \nonly column fitted is added.
", slug = "value"), list(title = "See also", contents = "

", slug = "see-also")),2021,list(name = "breathtestcore", version = "0.8.0.9001", authors = "Dieter Menne, Benjamin Misselwitz, Mark Fox, University Hospital of Zurich, Dep. Gastroenterology"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "breathtestcore"),list(css = "C:/Users/Dieter/Documents/RPackages/breathtestcore/pkgdown/extra.css"),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Reference\n
\n
\n  Articles\n
", right = "
\n  \n    \n     \n  \n
"),list(bootswatch = "cosmo", .present = TRUE),list(version = "1.6.1"),FALSE