Tabulates per-group breath test parameters

Given a fit to 13C breath test curves, computes absolute values and their conficence intervals of parameters, e.g. of the half emptying time t50. Generic S3 method for class breathtestfit.

coef_by_group(fit, ...)

Arguments

fit
Object of class breathtestfit, for example from nlme_fit, nls_fit or stan_fit

...
Not used

Value

A tibble of class coef_by_group with columns

parameter

Parameter of fit, e.g. beta, k, m, t50

method

Method used to compute parameter. exp_beta refers to primary fit parameters beta, k, m. maes_ghoos uses the method from Maes B D, Ghoos Y F, Rutgeerts P J, Hiele M I, Geypens B and Vantrappen G 1994 Dig. Dis. Sci. 39 S104-6. bluck_coward is the self-correcting method from Bluck L J C and Coward W A 2006 <doi:10.1088/0967-3334/27/3/006>

group

Grouping parameter of the fit, e.g. patient, normal, liquid, solid

estimate

Parameter estimate

conf.low, conf.high

Lower and upper 95 estimate.

diff_group

Letters a, b, c indicate that parameter would be in mutually significantly different groups. Letter combinations like ab or abc indicated that this parameter is not significantly different from the given other groups in a Tukey-corrected pairwise test.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

data("usz_13c")
data = usz_13c %>%
  dplyr::filter( patient_id %in%
    c("norm_001", "norm_002", "norm_003", "norm_004", "pat_001", "pat_002","pat_003")) %>%
  cleanup_data()
fit = nls_fit(data)
coef_by_group(fit)

#> |======================================================|100% ~0 s remaining     

#> # A tibble: 27 x 7
#>    parameter   method         group  estimate  conf.low conf.high diff_group
#>        <chr>    <chr>         <chr>     <dbl>     <dbl>     <dbl>      <chr>
#>  1      beta exp_beta liquid_normal  2.088000  1.345000   2.83100          a
#>  2      beta exp_beta       patient  2.657000  1.799000   3.51500          a
#>  3      beta exp_beta  solid_normal  2.204000  1.461000   2.94700          a
#>  4  deviance exp_beta liquid_normal  4.878000 -1.424000  11.18000         ab
#>  5  deviance exp_beta       patient 13.200000  5.926000  20.48000          b
#>  6  deviance exp_beta  solid_normal  3.406000 -2.897000   9.70800          a
#>  7         k exp_beta liquid_normal  0.011020  0.008603   0.01344          a
#>  8         k exp_beta       patient  0.008128  0.005335   0.01092          a
#>  9         k exp_beta  solid_normal  0.009920  0.007501   0.01234          a
#> 10         m exp_beta liquid_normal 26.840000 19.080000  34.60000          b
#> # ... with 17 more rows

fit = nlme_fit(data)
coef_by_group(fit)

#> # A tibble: 24 x 7
#>    parameter       method         group  estimate  conf.low conf.high
#>        <chr>        <chr>         <chr>     <dbl>     <dbl>     <dbl>
#>  1      beta     exp_beta liquid_normal  2.095000  1.688000   2.50100
#>  2      beta     exp_beta       patient  2.498000  2.029000   2.96800
#>  3      beta     exp_beta  solid_normal  2.238000  1.831000   2.64400
#>  4         k     exp_beta liquid_normal  0.011190  0.009325   0.01306
#>  5         k     exp_beta       patient  0.007847  0.005694   0.01000
#>  6         k     exp_beta  solid_normal  0.010070  0.008202   0.01193
#>  7         m     exp_beta liquid_normal 26.540000 20.380000  32.70000
#>  8         m     exp_beta       patient 23.960000 16.850000  31.08000
#>  9         m     exp_beta  solid_normal 17.460000 11.300000  23.62000
#> 10       t50 bluck_coward liquid_normal 34.730000 19.610000  49.85000
#> # ... with 14 more rows, and 1 more variables: diff_group <chr>