Tabulates breath test parameter differences of groups

Given a fit to 13C breath test curves, computes between-group confidence intervals and p-values, for examples of the half emptying time t50, with correction for multiple testing.

coef_diff_by_group(fit, mcp_group = "Tukey", reference_group = NULL, ...)

Arguments

fit
Object of class breathtestfit, for example from nlme_fit, nls_fit

mcp_group
"Tukey" (default) for all pairwise comparisons, "Dunnett" for comparisons relative to the reference group.

reference_group
Used as the first group and as reference group for mcp_group == "Dunnett"

...
Not used

Value

A tibble of class coef_diff_by_group with columns

parameter

Parameter of fit, e.g. beta, k, m, t50

method

Method used to compute parameter. exp_beta refers to primary fit parameters beta, k, m. maes_ghoos uses the method from Maes B D, Ghoos Y F, Rutgeerts P J, Hiele M I, Geypens B and Vantrappen G 1994 Dig. Dis. Sci. 39 S104-6. bluck_coward is the self-correcting method from Bluck L J C and Coward W A 2006 <doi:10.1088/0967-3334/27/3/006>

groups

Which pairwise difference, e.g solid - liquid

estimate

Estimate of the difference

conf.low, conf.high

Lower and upper 95 A comparison is significantly different from zero when both estimates have the same sign.

p.value

p-value of the difference against 0, corrected for multiple testing

Examples

library(dplyr)
data("usz_13c")
data = usz_13c %>%
  dplyr::filter( patient_id %in%
    c("norm_001", "norm_002", "norm_003", "norm_004", "pat_001", "pat_002","pat_003")) %>%
  cleanup_data()
fit = nls_fit(data)
coef_diff_by_group(fit)

#> # A tibble: 27 x 7
#>    parameter   method                       groups  estimate   conf.low
#>        <chr>    <chr>                        <chr>     <dbl>      <dbl>
#>  1      beta exp_beta      patient - liquid_normal  0.569000  -0.551600
#>  2      beta exp_beta solid_normal - liquid_normal  0.116200  -0.203200
#>  3      beta exp_beta       solid_normal - patient -0.452800  -1.573000
#>  4  deviance exp_beta      patient - liquid_normal  8.325000  -1.170000
#>  5  deviance exp_beta solid_normal - liquid_normal -1.473000  -4.543000
#>  6  deviance exp_beta       solid_normal - patient -9.798000 -19.290000
#>  7         k exp_beta      patient - liquid_normal -0.002894  -0.006516
#>  8         k exp_beta solid_normal - liquid_normal -0.001102  -0.003869
#>  9         k exp_beta       solid_normal - patient  0.001792  -0.001830
#> 10         m exp_beta      patient - liquid_normal -2.463000 -14.160000
#> # ... with 17 more rows, and 2 more variables: conf.high <dbl>, p.value <dbl>

fit = nlme_fit(data)
coef_diff_by_group(fit)

#> # A tibble: 24 x 7
#>    parameter       method                       groups  estimate   conf.low
#>        <chr>        <chr>                        <chr>     <dbl>      <dbl>
#>  1      beta     exp_beta      patient - liquid_normal  0.403600  -0.207100
#>  2      beta     exp_beta solid_normal - liquid_normal  0.142900  -0.131000
#>  3      beta     exp_beta       solid_normal - patient -0.260600  -0.871300
#>  4         k     exp_beta      patient - liquid_normal -0.003343  -0.006133
#>  5         k     exp_beta solid_normal - liquid_normal -0.001124  -0.003005
#>  6         k     exp_beta       solid_normal - patient  0.002220  -0.000570
#>  7         m     exp_beta      patient - liquid_normal -2.572000 -11.860000
#>  8         m     exp_beta solid_normal - liquid_normal -9.074000 -11.950000
#>  9         m     exp_beta       solid_normal - patient -6.501000 -15.790000
#> 10       t50 bluck_coward      patient - liquid_normal 31.940000   9.044000
#> # ... with 14 more rows, and 2 more variables: conf.high <dbl>, p.value <dbl>

# TODO: Add example for Stan fit typecast to class \code{breathtestfit} to compute
# confidence intervals instead of credible intervals