Broom-style tidying methods for breathtestfit

Broom-method tidy to streamline the results of class breathttestfit as generated by nls_fit or nlme_fit. Returns the fit coefficients and half-emptying time t50 with the Maes/Ghoos method; additional parameters should be extracted with coef.

# S3 method for breathtestfit
tidy(x, ...)

Arguments

x
Object of class breathttestfit

...
other parameters passed to methods

Value

A tibble/data frame with columns

patient_id

Patient Id (character)

group

Treatment or patient group (character)

m

Fraction metabolized

k

Time constant (1/minutes)

beta

The so-called lag parameters, no dimension

t50

Emptying half time in minutes as calculated following Maes/Ghoos

See also

tidy

Examples

library(broom)
# Generate simulated data
data = cleanup_data(simulate_breathtest_data()$data)
# Fit with the population method
fit = nlme_fit(data)
# Output coefficients
tidy(fit)

#> # A tibble: 10 x 6
#>    patient_id group     m      k  beta   t50
#>  *      <chr> <chr> <dbl>  <dbl> <dbl> <dbl>
#>  1     rec_01     A    31 0.0083   1.6   128
#>  2     rec_02     A    41 0.0068   1.8   164
#>  3     rec_03     A    40 0.0121   1.6    85
#>  4     rec_04     A    54 0.0105   1.9   111
#>  5     rec_05     A    16 0.0147   2.6    98
#>  6     rec_06     A    26 0.0121   2.3   110
#>  7     rec_07     A    50 0.0104   1.8   110
#>  8     rec_08     A    51 0.0081   2.7   183
#>  9     rec_09     A    49 0.0096   2.3   138
#> 10     rec_10     A    27 0.0105   2.0   115

# All coefficients in the long form
coef(fit)

#> # A tibble: 80 x 5
#>    patient_id group parameter                  method    value
#>         <chr> <chr>     <chr>                   <chr>    <dbl>
#>  1     rec_01     A         m                exp_beta  30.6445
#>  2     rec_01     A         k                exp_beta   0.0083
#>  3     rec_01     A      beta                exp_beta   1.6389
#>  4     rec_01     A       t50            bluck_coward  22.9120
#>  5     rec_01     A       t50              maes_ghoos 127.5256
#>  6     rec_01     A       t50 maes_ghoos_scintigraphy  54.8532
#>  7     rec_01     A      tlag            bluck_coward -23.8556
#>  8     rec_01     A      tlag              maes_ghoos  59.1786
#>  9     rec_02     A         m                exp_beta  41.1134
#> 10     rec_02     A         k                exp_beta   0.0068
#> # ... with 70 more rows