Broom-style tidying methods for breathtestfit

Broom-method tidy to streamline the results of class breathttestfit as generated by nls_fit or nlme_fit. Returns the fit coefficients and half-emptying time t50 with the Maes/Ghoos method; additional parameters should be extracted with coef.

# S3 method for breathtestfit
tidy(x, ...)

Arguments

x
Object of class breathttestfit

...
other parameters passed to methods

Value

A tibble/data frame with columns

patient_id

Patient Id (character)

group

Treatment or patient group (character)

m

Fraction metabolized

k

Time constant (1/minutes)

beta

The so-called lag parameters, no dimension

t50

Emptying half time in minutes as calculated following Maes/Ghoos

See also

Examples

library(broom)
# Generate simulated data
data = cleanup_data(simulate_breathtest_data()$data)
# Fit with the population method
fit = nlme_fit(data)
# Output coefficients
tidy(fit)

#> # A tibble: 10 x 6
#>    patient_id group     m       k  beta   t50
#>    <chr>      <chr> <dbl>   <dbl> <dbl> <dbl>
#>  1 rec_01     A      30.6 0.00835  1.64 128. 
#>  2 rec_02     A      41.1 0.00684  1.76 164. 
#>  3 rec_03     A      40.0 0.0121   1.58  85.2
#>  4 rec_04     A      54.3 0.0105   1.86 111. 
#>  5 rec_05     A      16.5 0.0147   2.56  97.7
#>  6 rec_06     A      26.1 0.0121   2.26 110. 
#>  7 rec_07     A      49.6 0.0104   1.82 110. 
#>  8 rec_08     A      50.8 0.00806  2.66 183. 
#>  9 rec_09     A      49.0 0.00964  2.26 138. 
#> 10 rec_10     A      27.4 0.0105   1.95 115. 

# All coefficients in the long form
coef(fit)

#> # A tibble: 80 x 5
#>    patient_id group parameter method                       value
#>    <chr>      <chr> <chr>     <chr>                        <dbl>
#>  1 rec_01     A     m         exp_beta                  30.6    
#>  2 rec_01     A     k         exp_beta                   0.00835
#>  3 rec_01     A     beta      exp_beta                   1.64   
#>  4 rec_01     A     t50       bluck_coward              22.9    
#>  5 rec_01     A     t50       maes_ghoos               128.     
#>  6 rec_01     A     t50       maes_ghoos_scintigraphy   54.9    
#>  7 rec_01     A     tlag      bluck_coward             -23.9    
#>  8 rec_01     A     tlag      maes_ghoos                59.2    
#>  9 rec_02     A     m         exp_beta                  41.1    
#> 10 rec_02     A     k         exp_beta                   0.00684
#> # ... with 70 more rows